ARQ-1220 Detect Jacoco JavaAgent and reuse coverage data report channel
When the client is running in Coverage Mode via the EclEmma Eclipse IDE
or the Maven Jacoco plugin, reuse the report channel used by the plugin instead of writing
to file.

This allow us to report the remote coverage data to the IDE when running
the "Coverage As -> JUnit" action.

Configuration of the destFile is done via the normal Maven Jacoco plugin.